refactor(api): migrate scoring service
