+++ api/lib/domain/services/placement-profile-service.js
@@ -2,6 +2,7 @@ import bluebird from 'bluebird';
 import _ from 'lodash';
 
 import { CertificationVersion } from '../../../src/certification/shared/domain/models/CertificationVersion.js';
+import * as scoringService from '../../../src/evaluation/domain/services/scoring/scoring-service.js';
 import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as assessmentResultRepository from '../../../src/shared/infrastructure/repositories/assessment-result-repository.js';
 import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
@@ -10,7 +11,6 @@ import * as knowledgeElementRepository from '../../infrastructure/repositories/k
 import * as knowledgeElementSnapshotRepository from '../../infrastructure/repositories/knowledge-element-snapshot-repository.js';
 import { PlacementProfile } from '../models/PlacementProfile.js';
 import { UserCompetence } from '../models/UserCompetence.js';
-import * as scoringService from './scoring/scoring-service.js';
 
 async function getPlacementProfile({
   userId,
