+++ api/tests/evaluation/unit/domain/services/scoring/scoring-service_test.js
@@ -2,9 +2,9 @@ import {
   MAX_REACHABLE_LEVEL,
   MAX_REACHABLE_PIX_BY_COMPETENCE,
   PIX_COUNT_BY_LEVEL,
-} from '../../../../../lib/domain/constants.js';
-import * as scoringService from '../../../../../lib/domain/services/scoring/scoring-service.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+} from '../../../../../../lib/domain/constants.js';
+import * as scoringService from '../../../../../../src/evaluation/domain/services/scoring/scoring-service.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Service | Scoring Service', function () {
   describe('#calculateScoringInformationForCompetence', function () {
